refactor: move `getPermissionsFromInputs` into its own file
